using Float64Array instead of BigUint64Array

+++ src/filter.ts
@@ -17,7 +17,7 @@ import { strToUint8, uint8ToStr } from "./comm";
  */
 const trzszMagicKeyPrefix = "::TRZSZ:TRANSFER:";
 const trzszMagicKeyRegExp = new RegExp(/::TRZSZ:TRANSFER:([SR]):(\d+\.\d+\.\d+)(:\d+)?/);
-const trzszMagicUint64 = new BigUint64Array(strToUint8(trzszMagicKeyPrefix).buffer, 0, 2);
+const trzszMagicUint64 = new Float64Array(strToUint8(trzszMagicKeyPrefix).buffer, 0, 2);
 
 /**
  * Find the trzsz magic key from output buffer.
@@ -45,7 +45,7 @@ export async function findTrzszMagicKey(output: string | ArrayBuffer | Uint8Arra
     if (idx < 0 || uint8.length - idx < 16) {
       return null;
     }
-    const uint64 = new BigUint64Array(uint8.buffer.slice(uint8.byteOffset + idx, uint8.byteOffset + idx + 16));
+    const uint64 = new Float64Array(uint8.buffer.slice(uint8.byteOffset + idx, uint8.byteOffset + idx + 16));
     if (uint64[0] == trzszMagicUint64[0] && uint64[1] == trzszMagicUint64[1]) {
       return uint8ToStr(uint8.subarray(idx));
     }
